requirements.txt
azure-storage-blob
geopandas
psycopg2
sqlalchemy
fastapi
uvicorn
twine
geoalchemy2
azure-functions